add defillama file and reformat
