+++ eagleproject/core/coingecko.py
@@ -5,7 +5,6 @@ from core.logging import get_logger
 log = get_logger(__name__)
 
 COINGECKO_ENDPOINT = "https://api.coingecko.com/api/v3"
-DEFILLAMA_ENDPOINT = "https://stablecoins.llama.fi"
 
 TICKER_TO_COINGECKO_ID = {
     "OGN": "origin-protocol",
@@ -60,25 +59,6 @@ def get_coin_history(ticker, from_timestamp, to_timestamp):
     if r.status_code != 200:
         raise Exception(
             "Failed to fetch ({}) volume data list from CoinGecko".format(
-                r.status_code
-            )
-        )
-
-    return r.json()
-
-def get_stablecoin_market_cap():
-    
-    uri = "{}/stablecoincharts/all".format(
-        DEFILLAMA_ENDPOINT,
-    )
-
-    log.debug("Fetching stablecoin market cap data from {}".format(uri))
-
-    r = requests.get(uri)
-
-    if r.status_code != 200:
-        raise Exception(
-            "Failed to fetch ({}) stablecoin market cap list from Defi Llama".format(
                 r.status_code
             )
         )
